Fix: 일기장 생성 시 메일 전송 분리
